Define Resource db label as BaseNode
Need to split helper functions out to prevent circular dependencies
