+++ src/common/db-label.decorator.ts
@@ -1,9 +1,4 @@
-import { uniq, without } from 'lodash';
-import { getParentTypes } from './parent-types';
-import { isResourceClass } from './resource.dto';
-import { AbstractClassType } from './types';
-
-const DbLabelSymbol = Symbol('DbLabelSymbol');
+export const DbLabelSymbol = Symbol('DbLabelSymbol');
 
 export const DbLabel =
   (...labels: string[]): PropertyDecorator & ClassDecorator =>
@@ -16,31 +11,3 @@ export const DbLabel =
     const current = Reflect.getMetadata(DbLabelSymbol, target, key) ?? [];
     Reflect.defineMetadata(DbLabelSymbol, [...current, ...labels], target, key);
   };
-
-export const getDbPropertyLabels = (
-  type: AbstractClassType<unknown>,
-  property: string
-) => {
-  const labels: string[] | undefined = Reflect.getMetadata(
-    DbLabelSymbol,
-    type.prototype,
-    property
-  );
-  const normalized = labels?.flatMap((l) => l.split(':'));
-  return uniq([...(normalized ?? []), 'Property']);
-};
-
-export const getDbClassLabels = (
-  type: AbstractClassType<unknown>
-): readonly string[] => {
-  const decorated: string[] | null = Reflect.getMetadata(DbLabelSymbol, type);
-  const labels =
-    decorated?.flatMap((l) => l.split(':')) ??
-    without(
-      getParentTypes(type)
-        .filter(isResourceClass)
-        .map((t) => t.name),
-      'Resource'
-    );
-  return uniq([...labels, 'BaseNode']);
-};
