+++ src/common/index.ts
@@ -3,7 +3,8 @@ export * from './buffer';
 export * from './temporal';
 export * from './context.type';
 export * from './date-filter.input';
-export * from './db-label.decorator';
+export { DbLabel } from './db-label.decorator';
+export * from './db-label.helpers';
 export * from './db-sort.decorator';
 export * from './db-unique.decorator';
 export * from './mutation-placeholder.output';
